+++ test/cypress.config.ts
@@ -3,9 +3,13 @@ import registerDurationMetricsPlugin from "../commonjs/plugin";
 
 export default defineConfig({
   e2e: {
-    "baseUrl": "https://example.cypress.io",
-    "video": false,
-    "screenshotOnRunFailure": false,
+    baseUrl: "https://example.cypress.io",
+    video: false,
+    screenshotOnRunFailure: false,
+    retries: {
+      runMode: 2,
+      openMode: 2,
+    },
     specPattern: 'cypress/integration/**/*.cy.{js,jsx,ts,tsx}',
     async setupNodeEvents(on, config) {
       registerDurationMetricsPlugin(on);